Move checksum-related things to their own module
Move ChecksumInfo and verify_checksum from package_managers.general to
core.checksum.
